+++ files/src/sdk/models/shared/fileresponse.ts
@@ -3,7 +3,7 @@
  */
 
 import { UserRoles, UserRoles$ } from "./userroles";
-import { z } from "zod";
+import * as z from "zod";
 
 export enum FileAccessScope {
     Private = "Private",