+++ files/RELEASES.md
@@ -29,3 +29,33 @@ Based on:
 - [typescript v2.0.1] files
 ### Releases
 - [NPM v2.0.1] https://www.npmjs.com/package/@wingspan/files/v/2.0.1 - files
+
+## 2024-03-11 01:23:55
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f51b81948601a18934c5
+- Speakeasy CLI 1.205.0 (2.279.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] files
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@wingspan/files/v/2.1.0 - files
+
+## 2024-03-18 01:24:06
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f51b81948601a18934c5
+- Speakeasy CLI 1.209.3 (2.281.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.1] files
+### Releases
+- [NPM v2.1.1] https://www.npmjs.com/package/@wingspan/files/v/2.1.1 - files
+
+## 2024-03-25 01:25:09
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f51b81948601a18934c5
+- Speakeasy CLI 1.221.0 (2.287.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.2] files
+### Releases
+- [NPM v2.1.2] https://www.npmjs.com/package/@wingspan/files/v/2.1.2 - files